make gql and rest play nice
